refactoring css files
