+++ src/styles/page.css
@@ -1,7 +1,4 @@
-.page {
-  display: flex;
-}
-.container {
+.page__container {
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
